Use a variable to track the channel order of the output buffer
The channel order of the output buffer is assumed to be the same as our
predefined order. It's hidden assumption when reading the code.
Introducing a variable stating the channel order of the output buffer
would improve the code readability. In addition, it would be able to set
any channel order of the output buffer by changing that variable.